Move ui operators to user_interface folder
